refactor: move auth redux files to store
