+++ src/actions/types.js
@@ -20,11 +20,3 @@ export const UPDATE_ITEM_ERROR = 'UPDATE_ITEM_ERROR'
 
 export const NEXT_WEEK = 'NEXT_WEEK'
 export const PREVIOUS_WEEK = 'PREVIOUS_WEEK'
-
-export const SIGN_IN_REQUEST = 'SIGN_IN_REQUEST'
-export const SIGN_IN_SUCCESS = 'SIGN_IN_SUCCESS'
-export const SIGN_IN_ERROR = 'SIGN_IN_ERROR'
-
-export const CURRENT_SESSION_REQUEST = 'CURRENT_SESSION_REQUEST'
-export const CURRENT_SESSION_SUCCESS = 'CURRENT_SESSION_SUCCESS'
-export const CURRENT_SESSION_ERROR = 'CURRENT_SESSION_ERROR'
