+++ src/scenes/login/index.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux'
 import LoginForm from './LoginForm'
-import { getCurrentSession, signIn } from '../../actions/authActions'
+import { getCurrentSession, signIn } from '../../store/auth/authActions'
 import { createLoadingSelector } from '../../selectors/loadingSelectors'
 import { createErrorMessageSelector } from '../../selectors/errorSelectors'
